fix(server): 404 on per-source sync — consolidate the duplicate endpoint
The new POST /topologies/:topologyId/sources/:sourceId/sync I added in
observations.ts was shadowed by a legacy handler at the same path in
topology-sources.ts (registered earlier on the same prefix). The legacy
one returned 404 because:
  - it interpreted :sourceId as the topology_data_sources.id (junction
    row id), but the UI sends the underlying data_source.id
  - its `service.get(sourceId)` lookup therefore failed → 404

Consolidate. The legacy handler (which used to do "syncOne actually
syncs ALL via the merge engine") is replaced in place with the
capability-dispatched per-source sync that:
  - resolves the attachment via `find(topologyId, dataSourceId, 'topology')`
  - autoscan-capable plugin → plugin.scan({seeds: []})
  - else topology-capable → plugin.fetchTopology(opts)
  - records the result as a `topology_observations` row
  - updates hysteresis, stamps lastSyncedAt for legacy surfaces
  - invalidates the parsed-topology cache so /render re-runs resolve()
  - never touches content_json (authored layer = editor-only)

The duplicate route in observations.ts is removed (left a one-line
comment pointing to the consolidated handler).

URL semantics are now uniform: `:sourceId` everywhere in the
topologies routes means data_source.id, not the junction-row id.

Validation: repo typecheck 34/34, svelte-check 0/0.
